// src/modules/powercard-balances/powercard-balances.interface.ts
export interface IPowercardBalances {
  StoreID: number;
  IsRegistered: boolean;
  CardNumber: number;
  Country: string;
  Status: number;
  GameChips: number;
  VideoChips: number;
  RewardChips: number;
  AttractionChips: number;
  Tickets: number;
  PointsToNextReward: number;
  EligibleRewardCount: number;
  CardEncoding: string;
}